Fix: Check for terminator in middle of first chunk
CRITICAL BUG: We were writing the entire first chunk without checking if
the terminator appeared in the middle of it. This caused us to send extra
data after the terminator, which would be the beginning of the NEXT
response. This corrupted the client's view and caused SABnzbd to report
missing articles.

Now we use find_terminator_end() on the first chunk just like we do for
all subsequent chunks, and only write up to the terminator position.

Fixes the 'missing articles' issue reported by SABnzbd.